+++ src/Pages/Dashboard/Components/TxHistory/Components/TxChart.tsx
@@ -8,6 +8,9 @@ import { formatDenom, isEmpty, subtractDays } from 'utils';
 import Big from 'big.js';
 import { useMediaQuery, useOrderbook } from '../../../../../redux/hooks';
 
+// To convert day of week
+const days = ['Sun', 'Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat'];
+
 const StyledChart = styled.div`
   height: 300px;
   width: 100%;
@@ -19,6 +22,15 @@ const StyledMessage = styled.div`
 
 const chartData = {
   color: [''],
+  title: {
+    text: '',
+    top: -5,
+    left: '5%',
+    textStyle: {
+      color: '',
+      fontWeight: 'normal',
+    },
+  },
   tooltip: {
     show: true,
     trigger: 'axis',
@@ -152,6 +164,10 @@ interface TxChartProps {
   txHistoryGran: string;
   span?: number;
   today?: Date;
+  showDayOfWeek?: boolean;
+  title?: string;
+  grid?: { left: string; right: string; bottom: string };
+  legendPadding?: number;
 }
 
 interface ParamsArray {
@@ -160,7 +176,16 @@ interface ParamsArray {
   data: { value: string; name: string };
 }
 
-export const TxChart = ({ txHistoryGran, data, span, today }: TxChartProps) => {
+export const TxChart = ({
+  txHistoryGran,
+  data,
+  span,
+  today,
+  showDayOfWeek = false,
+  title = '',
+  grid,
+  legendPadding = 20,
+}: TxChartProps) => {
   const [chart, setChart] = useState(null);
   const chartElementRef = useRef(null);
   const theme = useTheme();
@@ -180,54 +205,111 @@ export const TxChart = ({ txHistoryGran, data, span, today }: TxChartProps) => {
     const transactions: ValueProps[] = [];
     const fees: ValueProps[] = [];
     const dateArray: TxHistoryProps[] = [];
+    let xAxisData: string[] = [];
 
     // If the txHistoryCount is different, we want to ensure we populate
     // zeros for each day in the time span
-    if (txHistoryCount - 1 !== span) {
-      // First, build the date array
-      for (let i = Number(span); i >= 0; i--) {
-        dateArray.push({
-          date: new Date(format(subtractDays(today, i), 'yyyy-MM-dd')).getTime(),
-          feepayer: null,
-          txCount: 0,
-          feeAmountInBaseToken: 0,
-          gasWanted: 0,
-          gasUsed: 0,
-          feeAmountInToken: 0,
-          feesPaidInUsd: 0,
-          maxTokenPriceUsd: 0,
-          minTokenPriceUsd: 0,
-          avgTokenPriceUsd: 0,
-        });
-      }
-      // Then, fill dateArray with data that exists
-      dateArray.forEach((item, idx) => {
-        const index = data.findIndex((element) => element.date === item.date);
-        if (index !== -1) {
-          dateArray[idx] = data[index];
+    if (txHistoryGran === 'DAY') {
+      if (txHistoryCount - 1 !== span) {
+        // First, build the date array
+        for (let i = Number(span); i >= 0; i--) {
+          dateArray.push({
+            date: new Date(format(subtractDays(today, i), 'yyyy-MM-dd')).getTime(),
+            feepayer: null,
+            txCount: 0,
+            feeAmountInBaseToken: 0,
+            gasWanted: 0,
+            gasUsed: 0,
+            feeAmountInToken: 0,
+            feesPaidInUsd: 0,
+            maxTokenPriceUsd: 0,
+            minTokenPriceUsd: 0,
+            avgTokenPriceUsd: 0,
+          });
         }
-      });
-    }
-    // Populate series information
-    const xAxisData = (dateArray.length === 0 ? data : dateArray).map(
-      ({ txCount, date, feesPaidInUsd, feeAmountInToken }) => {
-        transactions.push({
-          value: txCount,
-          name: new Date(date).toISOString(),
+        // Then, fill dateArray with data that exists
+        dateArray.forEach((item, idx) => {
+          const index = data.findIndex((element) => element.date === item.date);
+          if (index !== -1) {
+            dateArray[idx] = data[index];
+          }
         });
-        fees.push({
-          value: feesPaidInUsd
-            ? Number(feesPaidInUsd)
-            : new Big(feeAmountInToken).times(currentPricing?.quote?.USD?.price || 0).toNumber(),
-          name: new Date(date).toISOString(),
-        });
-        return format(
-          parseISO(new Date(date).toISOString()),
-          granIsDay ? 'MMM dd' : granIsMonth ? 'MMM' : 'MM/dd, hh:mm'
-        );
       }
-    );
-
+      // Populate series information
+      xAxisData = (dateArray.length === 0 ? data : dateArray).map(
+        ({ txCount, date, feesPaidInUsd, feeAmountInToken }) => {
+          transactions.push({
+            value: txCount,
+            name: new Date(date).toISOString(),
+          });
+          fees.push({
+            value: feesPaidInUsd
+              ? Number(feesPaidInUsd)
+              : new Big(feeAmountInToken).times(currentPricing?.quote?.USD?.price || 0).toNumber(),
+            name: new Date(date).toISOString(),
+          });
+          return showDayOfWeek
+            ? days[new Date(date).getDay()]
+            : format(
+                parseISO(new Date(date).toISOString()),
+                granIsDay ? 'MMM dd' : granIsMonth ? 'MMM' : 'MM/dd, hh:mm'
+              );
+        }
+      );
+    } else {
+      // For yearly data
+      if (txHistoryCount - 1 !== 13) {
+        // First, build the date array
+        for (let i = 13; i >= 0; i--) {
+          // Here, the date needs to be 18:00 on the last day of the previous month
+          dateArray.push({
+            date: new Date(
+              new Date(
+                new Date(format(subtractDays(today, i * 30), 'yyyy-MM-dd')).setDate(1)
+              ).setHours(-6)
+            ).getTime(),
+            feepayer: null,
+            txCount: 0,
+            feeAmountInBaseToken: 0,
+            gasWanted: 0,
+            gasUsed: 0,
+            feeAmountInToken: 0,
+            feesPaidInUsd: 0,
+            maxTokenPriceUsd: 0,
+            minTokenPriceUsd: 0,
+            avgTokenPriceUsd: 0,
+          });
+        }
+        // Then, fill dateArray with data that exists
+        dateArray.forEach((item, idx) => {
+          const index = data.findIndex((element) => element.date === item.date);
+          if (index !== -1) {
+            dateArray[idx] = data[index];
+          }
+        });
+      }
+      // Populate series information
+      xAxisData = (dateArray.length === 0 ? data : dateArray).map(
+        ({ txCount, date, feesPaidInUsd, feeAmountInToken }) => {
+          transactions.push({
+            value: txCount,
+            name: new Date(date).toISOString(),
+          });
+          fees.push({
+            value: feesPaidInUsd
+              ? Number(feesPaidInUsd)
+              : new Big(feeAmountInToken).times(currentPricing?.quote?.USD?.price || 0).toNumber(),
+            name: new Date(date).toISOString(),
+          });
+          return showDayOfWeek
+            ? days[new Date(date).getDay()]
+            : format(
+                parseISO(new Date(date).toISOString()),
+                granIsDay ? 'MMM dd' : granIsMonth ? 'MMM' : 'MM/dd, hh:mm'
+              );
+        }
+      );
+    }
     // Now set chart data dynamically
     const colors: string[] = [theme.CHART_LINE_MAIN, theme.CHART_PIE_YES];
     chartData.color = colors;
@@ -296,6 +378,15 @@ export const TxChart = ({ txHistoryGran, data, span, today }: TxChartProps) => {
       )}</div> ${returnString}`;
       return returnString;
     };
+    // Set chart title info if provided
+    chartData.title.text = title;
+    chartData.title.textStyle.color = theme.FONT_PRIMARY;
+    // If different grid instructions provided, set them
+    if (grid) {
+      chartData.grid = grid;
+    }
+    // Set custom item gap value
+    chartData.legend.padding = legendPadding;
   }, [
     data,
     granIsDay,
@@ -307,6 +398,11 @@ export const TxChart = ({ txHistoryGran, data, span, today }: TxChartProps) => {
     span,
     today,
     txHistoryCount,
+    showDayOfWeek,
+    title,
+    grid,
+    legendPadding,
+    txHistoryGran,
   ]);
   // Legend
   chartData.legend.textStyle = {