+++ src/Pages/TxDash/Components/index.ts
@@ -1,0 +1,5 @@
+export * from './TxByHour';
+export * from './TxDaysOfWeek';
+export * from './TxHeatmap';
+export * from './TxHistory';
+export * from './TxNodeLink';