+++ src/redux/services/index.ts
@@ -2,4 +2,5 @@ export * from './accounts';
 export * from './chaincodeId';
 export * from './grants';
 export * from './serviceApi';
+export * from './txHeatmap';
 export * from './txs';